Add files via upload
The project should be able to run as is. It will only need user input once for the distance until its destination. A random number generator for the obstacle is also possible to be used in the code, but as is it is set up to repeat the simulation that outputted the plots in the paper.